python/tests/test_open_ai.py
from openai import OpenAI
import os
from dotenv import load_dotenv

load_dotenv()

key = os.getenv("OPENAI_API_KEY")
if not key:
    raise RuntimeError("OPENAI_API_KEY is missing. Add it to your .env file.")
client = OpenAI(api_key=key)

print("OK!")